+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Added `isPerfectScrollbar` prop to Modal.Body
 
 ## 0.1.0
 * Initial version